requirements.txt
fastapi
uvicorn
pillow
pytesseract
easyocr
pymongo
motor
python-multipart
loguru
pytest
